ecommerce-faq-chatbot/requirements.txt
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
peft>=0.6.0
bitsandbytes>=0.41.0
accelerate>=0.24.0
trl>=0.7.0
sentencepiece>=0.1.99
protobuf>=3.20.0
scipy>=1.11.0
nltk>=3.8.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
einops>=0.7.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
langchain>=0.0.300
streamlit>=1.28.0
